requirements.txt
absl-py==1.1.0
accelerate==0.11.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
cachetools==5.2.0
certifi==2022.6.15
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1696001773319/work
charset-normalizer==2.0.12
click==8.1.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
cycler==0.11.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1707444446795/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
Deprecated==1.2.13
docker-pycreds==0.4.0
einops==0.4.1
einops-exts==0.0.3
ema-pytorch==0.0.8
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1714422806336/work
fonttools==4.34.4
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
gudhi==3.9.0
h5py==3.7.0
huggingface_hub @ file:///home/conda/feedstock_root/build_artifacts/huggingface_hub_1711986612800/work
humanize==4.2.2
hydra-core==1.2.0
idna==3.3
imageio==2.19.3
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1708996548741/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1683289033986/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
joblib==1.1.0
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1699283905679/work
jupyter_core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1704727043329/work
kiwisolver==1.4.3
lxml==4.9.1
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
monai==1.3.0
multidict==6.0.2
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==2.8.5
nibabel==4.0.1
nilearn==0.9.1
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1655988925436/work
oauthlib==3.2.0
omegaconf==2.2.3
opencv-python==4.9.0.80
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==1.4.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pathtools==0.1.2
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.1.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1706713388748/work
promise==2.3
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.19.4
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1705722395397/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1711811537435/work
pyDeprecate==0.3.1
pydicom==2.3.0
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyparsing==3.1.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1709299778482/work
pytorch-lightning==1.6.4
pytz==2022.1
PyWavelets==1.3.0
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1695373436676/work
pyzmq==19.0.2
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rotary-embedding-torch==0.1.5
rsa==4.8
safetensors @ file:///home/conda/feedstock_root/build_artifacts/safetensors_1713253436573/work
scikit-image==0.19.3
scikit-learn==1.1.2
scikit-video==1.1.11
scipy==1.8.1
seaborn==0.11.2
sentry-sdk==1.7.2
setproctitle==1.2.3
shortuuid==1.0.9
SimpleITK==2.1.1.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.8.3
timm @ file:///home/conda/feedstock_root/build_artifacts/timm_1607370272046/work
toml==0.10.2
torch==1.12.1.post201
torch-tb-profiler==0.4.0
torch-two-sample==0.1
torchaudio==0.12.1
torchio==0.18.80
torchmetrics==0.9.1
torchvision @ file:///home/conda/feedstock_root/build_artifacts/torchvision-split_1658664666016/work
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1708363076094/work
tqdm==4.64.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1704212992681/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1708904622550/work
urllib3==1.26.9
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
